backend/requirements.txt
annotated-types
anyio
astroid
certifi
click
dill
dnspython
email-validator
environs
fastapi
fastapi_cors
h11
httpcore
httptools
httpx
idna
isort
itsdangerous
Jinja2
lazy-object-proxy
MarkupSafe
marshmallow
mccabe
orjson
packaging
platformdirs
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
pymongo
python-dotenv
python-multipart
PyYAML
sniffio
snowballstemmer
starlette
tomlkit
typing_extensions
ujson
uvicorn
uvloop
watchfiles
websockets
wrapt
